third_party/WebKit/LayoutTests/fast/dom/Range/range-toString-non-anchor-no-crash.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<script>
description('Tests that performing toString() over a Range without a root does not crash.');

var t1 = document.createElement("textarea");
var t2 = document.createElement("textarea");
t1.appendChild(document.createTextNode("PASS"));
t1.appendChild(t2);

var r = document.createRange();
r.setStart(t1, 0);
r.setEnd(t2, 0);

shouldBeEqualToString('r.toString()', 'PASS');
</script>
